+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Button colour to be set by using an explicit colour option.
 - Colour of the green buttons to be a bit darker.
+- Placeholder for some of the images to be blurred instead of single colour.
 
 ## [0.14.2] - 2021-03-27